backend/requirements.txt
Django>=4.0
djangorestframework
python-dateutil